Added homework info in homework popup
Homework info has been added to the calendar on the dashboard. Currently all homework for a course is shown due to eventID not being present for homework. Links dont work due to no unique way of going to them based on database info. If homework info isn't loaded then either click dirrectly, close the modal or refresh the page